feat: ProductService 에서 장비 통계 로직을 ReviewService로 이동
