+++ backend/src/test/java/com/woowacourse/f12/application/product/ProductServiceTest.java
@@ -1,18 +1,6 @@
 package com.woowacourse.f12.application.product;
 
-import static com.woowacourse.f12.domain.member.CareerLevel.JUNIOR;
-import static com.woowacourse.f12.domain.member.CareerLevel.MID_LEVEL;
-import static com.woowacourse.f12.domain.member.JobType.BACKEND;
-import static com.woowacourse.f12.domain.member.JobType.ETC;
 import static com.woowacourse.f12.domain.product.Category.KEYBOARD;
-import static com.woowacourse.f12.presentation.member.CareerLevelConstant.JUNIOR_CONSTANT;
-import static com.woowacourse.f12.presentation.member.CareerLevelConstant.MID_LEVEL_CONSTANT;
-import static com.woowacourse.f12.presentation.member.CareerLevelConstant.NONE_CONSTANT;
-import static com.woowacourse.f12.presentation.member.CareerLevelConstant.SENIOR_CONSTANT;
-import static com.woowacourse.f12.presentation.member.JobTypeConstant.BACKEND_CONSTANT;
-import static com.woowacourse.f12.presentation.member.JobTypeConstant.ETC_CONSTANT;
-import static com.woowacourse.f12.presentation.member.JobTypeConstant.FRONTEND_CONSTANT;
-import static com.woowacourse.f12.presentation.member.JobTypeConstant.MOBILE_CONSTANT;
 import static com.woowacourse.f12.presentation.product.CategoryConstant.KEYBOARD_CONSTANT;
 import static com.woowacourse.f12.presentation.product.CategoryConstant.MOUSE_CONSTANT;
 import static com.woowacourse.f12.support.fixture.ProductFixture.KEYBOARD_1;
@@ -30,18 +18,14 @@ import static org.mockito.Mockito.verify;
 import com.woowacourse.f12.domain.inventoryproduct.InventoryProductRepository;
 import com.woowacourse.f12.domain.product.Product;
 import com.woowacourse.f12.domain.product.ProductRepository;
-import com.woowacourse.f12.domain.review.CareerLevelCount;
-import com.woowacourse.f12.domain.review.JobTypeCount;
 import com.woowacourse.f12.domain.review.ReviewRepository;
 import com.woowacourse.f12.dto.request.product.ProductCreateRequest;
 import com.woowacourse.f12.dto.request.product.ProductSearchRequest;
 import com.woowacourse.f12.dto.request.product.ProductUpdateRequest;
 import com.woowacourse.f12.dto.response.product.ProductPageResponse;
 import com.woowacourse.f12.dto.response.product.ProductResponse;
-import com.woowacourse.f12.dto.response.product.ProductStatisticsResponse;
 import com.woowacourse.f12.exception.notfound.ProductNotFoundException;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -119,34 +103,6 @@ class ProductServiceTest {
                 () -> assertThatThrownBy(() -> productService.findById(1L))
                         .isExactlyInstanceOf(ProductNotFoundException.class),
                 () -> verify(productRepository).findById(1L)
-        );
-    }
-
-    @Test
-    void 특정_제품의_사용자의_연차와_직군의_비율을_반환한다() {
-        // given
-        Long productId = 1L;
-        given(productRepository.existsById(productId))
-                .willReturn(true);
-        given(reviewRepository.findCareerLevelCountByProductId(productId))
-                .willReturn(List.of(new CareerLevelCount(JUNIOR, 1), new CareerLevelCount(MID_LEVEL, 1)));
-        given(reviewRepository.findJobTypeCountByProductId(productId))
-                .willReturn(List.of(new JobTypeCount(BACKEND, 1), new JobTypeCount(ETC, 1)));
-
-        // when
-        ProductStatisticsResponse productStatisticsResponse = productService.calculateMemberStatisticsById(productId);
-
-        // then
-        assertAll(
-                () -> verify(productRepository).existsById(productId),
-                () -> verify(reviewRepository).findCareerLevelCountByProductId(productId),
-                () -> verify(reviewRepository).findJobTypeCountByProductId(productId),
-                () -> assertThat(productStatisticsResponse.getCareerLevel()).usingRecursiveComparison()
-                        .isEqualTo(Map.of(NONE_CONSTANT, 0.0, JUNIOR_CONSTANT, 0.5, MID_LEVEL_CONSTANT, 0.5,
-                                SENIOR_CONSTANT, 0.0)),
-                () -> assertThat(productStatisticsResponse.getJobType()).usingRecursiveComparison()
-                        .isEqualTo(Map.of(FRONTEND_CONSTANT, 0.0, BACKEND_CONSTANT, 0.5, MOBILE_CONSTANT, 0.0,
-                                ETC_CONSTANT, 0.5))
         );
     }
