+++ backend/src/test/java/com/woowacourse/f12/presentation/product/ProductControllerTest.java
@@ -1,13 +1,5 @@
 package com.woowacourse.f12.presentation.product;
 
-import static com.woowacourse.f12.domain.member.CareerLevel.JUNIOR;
-import static com.woowacourse.f12.domain.member.CareerLevel.MID_LEVEL;
-import static com.woowacourse.f12.domain.member.CareerLevel.NONE;
-import static com.woowacourse.f12.domain.member.CareerLevel.SENIOR;
-import static com.woowacourse.f12.domain.member.JobType.BACKEND;
-import static com.woowacourse.f12.domain.member.JobType.ETC;
-import static com.woowacourse.f12.domain.member.JobType.FRONTEND;
-import static com.woowacourse.f12.domain.member.JobType.MOBILE;
 import static com.woowacourse.f12.exception.ErrorCode.EXPIRED_ACCESS_TOKEN;
 import static com.woowacourse.f12.exception.ErrorCode.INVALID_PAGING_PARAM;
 import static com.woowacourse.f12.exception.ErrorCode.INVALID_REQUEST_BODY_TYPE;
@@ -43,8 +35,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.woowacourse.f12.application.auth.token.JwtProvider;
 import com.woowacourse.f12.application.auth.token.MemberPayload;
 import com.woowacourse.f12.application.product.ProductService;
-import com.woowacourse.f12.domain.member.CareerLevel;
-import com.woowacourse.f12.domain.member.JobType;
 import com.woowacourse.f12.domain.member.Role;
 import com.woowacourse.f12.domain.product.Product;
 import com.woowacourse.f12.dto.request.product.ProductCreateRequest;
@@ -53,12 +43,10 @@ import com.woowacourse.f12.dto.request.product.ProductUpdateRequest;
 import com.woowacourse.f12.dto.response.PopularProductsResponse;
 import com.woowacourse.f12.dto.response.product.ProductPageResponse;
 import com.woowacourse.f12.dto.response.product.ProductResponse;
-import com.woowacourse.f12.dto.response.product.ProductStatisticsResponse;
 import com.woowacourse.f12.exception.notfound.ProductNotFoundException;
 import com.woowacourse.f12.presentation.PresentationTest;
 import com.woowacourse.f12.support.ErrorCodeSnippet;
 import java.util.List;
-import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
@@ -194,45 +182,6 @@ class ProductControllerTest extends PresentationTest {
                 .andDo(print());
 
         verify(productService).findById(0L);
-    }
-
-    @Test
-    void 특정_제품에_대한_사용자_통계_조회_성공() throws Exception {
-        // given
-        Map<CareerLevel, Double> careerLevel = Map.of(NONE, 0.0, JUNIOR, 0.5,
-                MID_LEVEL, 0.0, SENIOR, 0.5);
-        Map<JobType, Double> jobType = Map.of(FRONTEND, 0.33, BACKEND,
-                0.33, MOBILE, 0.33, ETC, 0.0);
-        given(productService.calculateMemberStatisticsById(anyLong()))
-                .willReturn(ProductStatisticsResponse.of(careerLevel, jobType));
-
-        // when
-        ResultActions resultActions = mockMvc.perform(
-                get("/api/v1/products/1/statistics")
-        );
-
-        // then
-        resultActions.andExpect(status().isOk())
-                .andDo(document("products-member-statistics-get", new ErrorCodeSnippet(PRODUCT_NOT_FOUND)))
-                .andDo(print());
-
-        verify(productService).calculateMemberStatisticsById(1L);
-    }
-
-    @Test
-    void 특정_제품에_대한_사용자_통계_조회_실패_제품이_존재하지_않을_경우() throws Exception {
-        // given
-        given(productService.calculateMemberStatisticsById(anyLong()))
-                .willThrow(new ProductNotFoundException());
-
-        // when
-        ResultActions resultActions = mockMvc.perform(get("/api/v1/products/1/statistics"));
-
-        // then
-        resultActions.andExpect(status().isNotFound())
-                .andDo(print());
-
-        verify(productService).calculateMemberStatisticsById(1L);
     }
 
     @Test
